// apps/web/src/routes/_dashboard/$user/repos/$repo/-components/__generated__/RepoGeneralInfo_repository.graphql.ts
export type RepoGeneralInfo_repository$data = {
  readonly description: string | null | undefined;
  readonly diskUsage: number | null | undefined;
  readonly forkCount: number;
  readonly hasDiscussionsEnabled: boolean;
  readonly hasIssuesEnabled: boolean;
  readonly hasProjectsEnabled: boolean;
  readonly hasWikiEnabled: boolean;
  readonly homepageUrl: string | null | undefined;
  readonly id: string;
  readonly isArchived: boolean;
  readonly isDisabled: boolean;
  readonly isFork: boolean;
  readonly isLocked: boolean;
  readonly isPrivate: boolean;
  readonly isTemplate: boolean;
  readonly isUserConfigurationRepository: boolean;
  readonly languages: {
    readonly edges: ReadonlyArray<{
      readonly node: {
        readonly color: string | null | undefined;
        readonly id: string;
        readonly name: string;
      };
      readonly size: number;
    } | null | undefined> | null | undefined;
    readonly totalSize: number;
  } | null | undefined;
  readonly name: string;
  readonly nameWithOwner: string;
  readonly openGraphImageUrl: string;
  readonly pushedAt: string | null | undefined;
  readonly repositoryTopics: {
    readonly nodes: ReadonlyArray<{
      readonly id: string;
      readonly topic: {
        readonly name: string;
      };
    } | null | undefined> | null | undefined;
  };
  readonly stargazerCount: number;
  readonly updatedAt: string;
  readonly url: string;
  readonly viewerHasStarred: boolean;
  readonly " $fragmentType": "RepoGeneralInfo_repository";
};
export type RepoGeneralInfo_repository$key = {
  readonly " $data"?: RepoGeneralInfo_repository$data;
  readonly " $fragmentSpreads": FragmentRefs<"RepoGeneralInfo_repository">;
};